conceito c, adicionar paginas de tag

+++ notes/views.py
@@ -16,18 +16,25 @@ def deleteNote(identif):
 def index(request):
     if request.method == 'POST':
         noteid = request.POST.get('edit_note')
+        deleteid = request.POST.get('delete_note_id')
         title = request.POST.get('titulo')
         content = request.POST.get('detalhes')
-        tag = request.POST.get('tags')
 
-        deleteid = request.POST.get('delete_id')
+        #Check if tag is None, and if i"t has a '#'
+        if request.POST.get('tags') == None:
+            tag = request.POST.get('tags')
+        else:
+            if request.POST.get('tags')[0] == "#":
+                tag = request.POST.get('tags').lower()
+            else:
+                tag = "#" + request.POST.get('tags').lower()    
 
         #Delete note
-        if deleteid != None:
+        if deleteid != None and deleteid != "None" and deleteid != '':
             deleteNote(deleteid)
         
         #Edit note
-        elif noteid != None:
+        elif noteid != None and noteid != "None" and noteid != '':
             editNote(noteid, title, content, tag)
         
         #Create note
